Moved philu_features app to openedx/features
