+++ lms/envs/common.py
@@ -2229,7 +2229,7 @@ INSTALLED_APPS = (
     'lms.djangoapps.philu_overrides',
 
     # Philu Features
-    'lms.djangoapps.philu_features',
+    'openedx.features.philu_features',
 
     #third party surveys
     'lms.djangoapps.third_party_surveys',
