+++ openedx/features/philu_features/urls.py
@@ -3,7 +3,7 @@ The urls for philu features app.
 """
 from django.conf.urls import url
 
-from lms.djangoapps.philu_features import views
+from openedx.features.philu_features import views
 
 urlpatterns = [
     url(r"^certificates/$", views.student_certificates, name="certificates"),
